requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
Django==5.0.4
djangorestframework==3.15.1
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
PySocks==1.7.1
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
Unidecode==1.3.8
urllib3==2.2.1
wsproto==1.2.0
